requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
pydub==0.25.1
python-dotenv==1.0.1
supabase==2.15.0
yt-dlp==2025.2.19
